+++ TimeFliesBy.lua
@@ -3,8 +3,10 @@ local _, tfb = ...
 local initialTimeChecked = false
 local WunderBar = tfb.WunderBar
 
+local specialBarVisibleTime = 60
+
 local alternativeWatch = 0
-local function updateMaxLvlBar(event)
+local function updateMaxLvlBar(event, message)
   if event == "UPDATE_FACTION" then
     local faction = tfb.reputation:GetReputationChange()
     if faction ~= nil then
@@ -12,7 +14,18 @@ local function updateMaxLvlBar(event)
       WunderBar:SetValues(faction.max, faction.current)
       local perc = floor((faction.current / faction.max) * 100)
       WunderBar:SetText(string.format("%s - %d / %d (%d%%)", faction.name, faction.current, faction.max, perc), faction.standing)
-      alternativeWatch = time() + 30
+      alternativeWatch = time() + specialBarVisibleTime
+    end
+  end
+
+  if event == "CHAT_MSG_SKILL" then
+    local skill = tfb.skill:ParseSkillMsg(message)
+    if skill ~= nil then
+      WunderBar:SetBar1Color(skill.GetColor())
+      WunderBar:SetValues(skill.max, skill.current)
+      local perc = floor((skill.current / skill.max) * 100)
+      WunderBar:SetText(string.format("%s - %d / %d (%d%%)", skill.name, skill.current, skill.max, perc))
+      alternativeWatch = time() + specialBarVisibleTime
     end
   end
 
@@ -29,6 +42,7 @@ local function initMaxLvlBar()
   C_Timer.NewTicker(5, updateMaxLvlBar)
   tfb.reputation:GetReputationChange() -- init
   tfb.events:Register("UPDATE_FACTION", "maxLvlBar", updateMaxLvlBar)
+  tfb.events:Register("CHAT_MSG_SKILL", "maxLvlBar", updateMaxLvlBar)
 end
 
 local sessionStart, sessionStartExp